app/api/requirements.txt
fastapi==0.108.0
uvicorn==0.25.0
onnxruntime==1.21.0
numpy==1.26.2
python-multipart==0.0.6
Pillow==10.1.0